+++ src/components/StackedAreaChart/components/StackedAreas/StackedAreas.tsx
@@ -1,30 +1,17 @@
 import React, {useMemo} from 'react';
 import isEqual from 'fast-deep-equal';
-import {animated, useSpring} from '@react-spring/web';
-import {area, line, Series} from 'd3-shape';
+import {area, line} from 'd3-shape';
 import type {ScaleLinear} from 'd3-scale';
-import type {Color, GradientStop} from 'types';
 
-import {LinearGradient} from '../../../LinearGradient';
-import {
-  curveStepRounded,
-  isGradientType,
-  uniqueId,
-} from '../../../../utilities';
+import type {Color} from '../../../../types';
+import {curveStepRounded, uniqueId} from '../../../../utilities';
 import {usePrevious, useTheme} from '../../../../hooks';
+import type {StackedSeries} from '../../types';
+import {Area} from '..';
 
 import styles from './StackedAreas.scss';
 
-type StackedSeries = Series<
-  {
-    [key: string]: number;
-  },
-  string
->;
-
 interface Props {
-  width: number;
-  height: number;
   transform: string;
   colors: Color[];
   stackedValues: StackedSeries[];
@@ -34,9 +21,12 @@ interface Props {
   theme?: string;
 }
 
+const FAST_DURATION = 100;
+const SLOW_DURATION = 325;
+const FAST_COUNT = 10;
+const DURATION_STEP = 25;
+
 export function Areas({
-  width,
-  height,
   stackedValues,
   transform,
   xScale,
@@ -49,17 +39,9 @@ export function Areas({
   const prevstackedValues = usePrevious(stackedValues);
   const valuesHaveNotUpdated = isEqual(prevstackedValues, stackedValues);
 
-  const spring = useSpring({
-    config: {duration: 1000},
-    width,
-    from: {
-      width: 0,
-    },
-    immediate: !isAnimated || valuesHaveNotUpdated,
-    reset: true,
-  });
+  const id = useMemo(() => uniqueId('stackedAreas'), []);
 
-  const areaShape = area<number[]>()
+  const areaGenerator = area<number[]>()
     .defined(
       ([firstPoint, lastPoint]) => !isNaN(firstPoint) && !isNaN(lastPoint),
     )
@@ -67,7 +49,7 @@ export function Areas({
     .y0(([firstPoint]) => yScale(firstPoint))
     .y1(([, lastPoint]) => yScale(lastPoint));
 
-  const lineShape = line<number[]>()
+  const lineGenerator = line<number[]>()
     .defined(
       ([firstPoint, lastPoint]) => !isNaN(firstPoint) && !isNaN(lastPoint),
     )
@@ -75,71 +57,40 @@ export function Areas({
     .y(([, lastPoint]) => yScale(lastPoint));
 
   if (selectedTheme.line.hasSpline) {
-    areaShape.curve(curveStepRounded);
-    lineShape.curve(curveStepRounded);
+    areaGenerator.curve(curveStepRounded);
+    lineGenerator.curve(curveStepRounded);
   }
 
-  const id = useMemo(() => uniqueId('stackedAreas'), []);
+  const duration = useMemo(() => {
+    const count = stackedValues.length;
+    const duration =
+      count > FAST_COUNT
+        ? FAST_DURATION
+        : SLOW_DURATION - count * DURATION_STEP;
+
+    return Math.round(duration / DURATION_STEP) * DURATION_STEP;
+  }, [stackedValues.length]);
 
   return (
-    <React.Fragment>
-      <clipPath id={id}>
-        <animated.rect
-          width={spring.width}
-          height={height + selectedTheme.line.width}
-          y={-selectedTheme.line.width}
-          fill="none"
-        />
-      </clipPath>
-      <g
-        transform={transform}
-        clipPath={`url(#${id})`}
-        className={styles.Group}
-      >
-        {stackedValues.map((value, index) => {
-          const shape = areaShape(value);
-          const line = lineShape(value);
-
-          if (shape == null || line == null) {
-            return null;
-          }
-
-          const currentColor = colors[index];
-          const isGradient = isGradientType(currentColor);
-
-          const gradient = isGradient
-            ? currentColor
-            : [{offset: 0, color: currentColor}];
-
-          return (
-            <React.Fragment key={index}>
-              <defs>
-                <LinearGradient
-                  id={`area-${id}-${index}`}
-                  gradient={gradient as GradientStop[]}
-                  gradientUnits="userSpaceOnUse"
-                  y1="100%"
-                  y2="0%"
-                />
-              </defs>
-              <path
-                key={`line-${index}`}
-                d={line}
-                fill="none"
-                stroke={`url(#area-${id}-${index})`}
-                strokeWidth={selectedTheme.line.width}
-              />
-              <path
-                key={index}
-                d={shape}
-                fill={`url(#area-${id}-${index})`}
-                fillOpacity="0.25"
-              />
-            </React.Fragment>
-          );
-        })}
-      </g>
-    </React.Fragment>
+    <g transform={transform} className={styles.Group}>
+      {stackedValues.map((data, index) => {
+        return (
+          <Area
+            animationIndex={stackedValues.length - 1 - index}
+            areaGenerator={areaGenerator}
+            colors={colors}
+            data={data}
+            duration={duration}
+            id={id}
+            index={index}
+            isImmediate={!isAnimated || valuesHaveNotUpdated}
+            key={index}
+            lineGenerator={lineGenerator}
+            selectedTheme={selectedTheme}
+          />
+        );
+      })}
+    </g>
   );
 }