+++ src/components/StackedAreaChart/components/index.ts
@@ -1,1 +1,2 @@
 export {StackedAreas} from './StackedAreas';
+export {Area} from './Area';